bestProduct operation fixed

+++ GreenChoice.Domain/Dtos/CommentReponseDto.cs
@@ -8,5 +8,5 @@ public class CommentReponseDto
     public string Text { get; set; }
     public string UserName { get; set; }
     public string ProductName { get; set; }
-    public int CommentScore { get; set; }
+    public float CommentScore { get; set; }
 }
